// src/app/models/mascotas-extraviadas.ts
export interface MascotasExtraviadas {
    id_mascota_extraviada?: number;
    nombre: string;
    raza: string;
    edad: number;
    sexo: string;
    especie: string;
    telefono: string;
    direccion: string;
    usuario_id: number;
    descripcion: string;
    fecha_extravio: string;
    fecha_publicacion: string;
    imagen?: string;
}